\name{HGNChelper-package}
\alias{HGNChelper-package}
\docType{package}
\title{Handy functions for working with HGNC gene symbols and
Affymetrix probeset identifiers.}
\description{Contains functions for
identifying and correcting HGNC gene symbols which have been converted
to date format by Excel, for reversibly converting between HGNC
symbols and valid R names, identifying invalid HGNC symbols and
correcting synonyms and outdated symbols which can be mapped to an
official symbol.}
\details{
\tabular{ll}{Package: \tab HGNChelper\cr
Authors: \tab Levi Waldron and Markus Riester\cr
Maintainer: \tab Levi Waldron <lwaldron.research@gmail.com>\cr
Depends: \tab R (>= 3.5.0)\cr
Version: \tab \Sexpr{packageDescription("HGNChelper")$Version}\cr
Date: \tab \Sexpr{packageDescription("HGNChelper")$Date}\cr
License: \tab GPL (>2.0)\cr
Title: \tab Handy functions for working with HGNC gene symbols and
Affymetrix probeset identifiers.\cr
URL: \tab https://waldronlab.io/HGNChelper/\cr
BugReports: \tab https://github.com/waldronlab/hgnchelper/issues \cr}
}
\author{Levi Waldron and Markus Riester}

\keyword{ package }


