% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affyToR.R
\name{affyToR}
\alias{affyToR}
\title{Title function to convert Affymetrix probeset identifiers to valid R names}
\usage{
affyToR(x)
}
\arguments{
\item{x}{vector of Affymetrix probeset identifiers, or any identifier which may with a digit.}
}
\value{
a character vector that is simply x with "affy." prepended to each value.
}
\description{
This function simply prepends "affy." to the probeset IDs to create
valid R names.  Reverse operation is done by the \link{rToAffy}
function.
}
