% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgnc.table}
\alias{hgnc.table}
\title{All current and withdrawn HGNC gene symbols and Excel mogrifications.}
\format{An object of class \code{data.frame} with 98431 rows and 2 columns.}
\source{
Extracted from 
\url{ftp://ftp.ebi.ac.uk/pub/databases/genenames/new/tsv/hgnc_complete_set.txt}
 and 
system.file("extdata/mog_map.csv", package="HGNChelper")
}
\usage{
hgnc.table
}
\description{
A dataframe with the first column providing a gene symbol or known alias
(including withdrawn symbols), second column providing the approved HGNC 
human gene symbol.
}
\details{
\itemize{
  \item Symbol. All valid, Excel-mogrified, and withdrawn symbols
  \item Approved.Symbol. Approved symbols.
}
}
\examples{
data("hgnc.table", package="HGNChelper")
head(hgnc.table)
}
\keyword{datasets}
