% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNChelper-package.R
\docType{package}
\name{HGNChelper-package}
\alias{HGNChelper}
\alias{HGNChelper-package}
\title{HGNChelper: Identify and Correct Invalid HGNC Human Gene Symbols and MGI Mouse Gene Symbols}
\description{
Contains functions for identifying and correcting HGNC human gene symbols and MGI mouse gene symbols which have been converted to date format by Excel, withdrawn, or aliased. Also contains functions for reversibly converting between HGNC symbols and valid R names.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/waldronlab/HGNChelper}
  \item \url{https://waldronlab.io/HGNChelper/}
  \item Report bugs at \url{https://github.com/waldronlab/HGNChelper/issues}
}

}
\author{
\strong{Maintainer}: Levi Waldron \email{lwaldron.research@gmail.com}

Authors:
\itemize{
  \item Sehyun Oh \email{sehyun.oh@sph.cuny.edu}
  \item Ayush Aggarwal \email{ayush.aggarwal@igib.in}
  \item Markus Riester \email{Markus.riester@novartis.com}
}

}
\keyword{internal}
