\name{HGNChelper-package}
\alias{HGNChelper-package}
\docType{package}
\title{Handy functions for working with HGNC gene symbols and Affymetrix probeset identifiers.}
\description{Contains functions for identifying and correcting HGNC gene symbols which have been converted to date format by Excel, and for reversibly converting between HGNC symbols and valid R names.}
\details{
\tabular{ll}{
Package: \tab HGNChelper\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}

\author{Levi Waldron}

\keyword{ package }


