\name{HGNChelper-package}
\alias{HGNChelper-package}
\docType{package}
\title{Handy functions for working with HGNC gene symbols and
Affymetrix probeset identifiers.}
\description{Contains functions for
identifying and correcting HGNC gene symbols which have been converted
to date format by Excel, for reversibly converting between HGNC
symbols and valid R names, identifying invalid HGNC symbols and
correcting synonyms and outdated symbols which can be mapped to an
official symbol.}
\details{
\tabular{ll}{Package: \tab HGNChelper\cr
Maintainer: \tab Levi Waldron <lwaldron@hsph.harvard.edu>\cr
Depends: \tab R (>= 2.10)\cr
Author: \tab Levi Waldron and Markus Riester\cr
Version: \tab \Sexpr{packageDescription("HGNChelper")$Version}\cr
Date: \tab \Sexpr{packageDescription("HGNChelper")$Date}\cr
License: \tab GPL (>2.0)\cr
Title: \tab Handy functions for working with HGNC gene symbols and
Affymetrix probeset identifiers.\cr
URL: \tab https://bitbucket.org/lwaldron/hgnchelper\cr
BugReports: \tab https://bitbucket.org/lwaldron/hgnchelper\cr}
}
\author{Levi Waldron and Markus Riester}

\keyword{ package }


