% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.table}
\alias{mouse.table}
\title{All current and withdrawn MGI mouse symbols and Excel mogrifications.}
\format{An object of class \code{data.frame} with 134371 rows and 2 columns.}
\source{
Extracted from \url{http://www.informatics.jax.org/downloads/reports/MGI_EntrezGene.rpt} and 
system.file("extdata/HGNChelper_mog_map_MGI_AMC_2016_03_30.csv", package="HGNChelper")
}
\usage{
mouse.table
}
\description{
A dataframe with the first column providing a gene symbol or known alias
(including withdrawn symbols), second column providing the approved MOUSE symbol.
}
\details{
\itemize{
  \item Symbol. All valid, Excel-mogrified, and withdrawn symbols
  \item Approved.Symbol. Approved symbols.
}
}
\examples{
data("mouse.table", package="HGNChelper")
head(mouse.table)
}
\keyword{datasets}
