\name{rToAffy}
\alias{rToAffy}
\title{function to convert the output of affyToR back to the original Affymetrix probeset identifiers.}
\description{This function simply strips the "affy." added by the \link{affyToR} function.}
\usage{rToAffy(x)}
\arguments{
  \item{x}{the character vector returned by the affyToR function.}
}

\value{a character vector of Affymetrix probeset identifiers.}

\author{Levi Waldron and Markus Riester}





