% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rToAffy.R
\name{rToAffy}
\alias{rToAffy}
\title{Title function to convert the output of affyToR back to the original Affymetrix probeset identifiers.}
\usage{
rToAffy(x)
}
\arguments{
\item{x}{the character vector returned by the affyToR function.}
}
\value{
a character vector of Affymetrix probeset identifiers.
}
\description{
This function simply strips the "affy." added by the \link{affyToR} function.
}
