% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rToSymbol.R
\name{rToSymbol}
\alias{rToSymbol}
\title{Title function to reverse the conversion made by symbolToR}
\usage{
rToSymbol(x)
}
\arguments{
\item{x}{the character vector returned by the symbolToR function.}
}
\value{
a character vector of HGNC gene symbols, which are not in general valid R names.
}
\description{
This function reverses the actions of the symbolToR function.
}
\seealso{
\code{\link{symbolToR}}
}
