% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolToR.R
\name{symbolToR}
\alias{symbolToR}
\title{Title function to *reversibly* convert HGNC gene symbols to valid R names.}
\usage{
symbolToR(x)
}
\arguments{
\item{x}{vector of HGNC symbols}
}
\value{
a vector of valid R names, of the same length as x, which can be
converted to the same HGNC symbols using the rToSymbol function.
}
\description{
This function reversibly converts HGNC gene symbols to valid R
names by prepending "symbol.", and making the following
substitutions: "-" to "hyphen", "@" to "ampersand", and "/" to
"forwardslash".
}
\seealso{
\code{\link{rToSymbol}}
}
